[level 3] Title: 입국심사, Time: 0.01 ms, Memory: 4.17 MB -BaekjoonHub

+++ 프로그래머스/lv3/43238. 입국심사/입국심사.cpp
@@ -20,7 +20,7 @@ long long solution(int n, vector<int> times)
         cnt = 0;
         
         for(int i = 0; i < times.size(); ++i)
-            cnt += midTime / times[i];
+            cnt += midTime / (long long)times[i];
         
         if(cnt < n) minTime = midTime + 1;
         else
